Move constraint* libraries out of the libraries/pacemaker directory otherwise compilation of recipe fails
